Add AskAISearchDialog component
This commit introduces a new AskAISearchDialog component that can be invoked via a UI button. The component uses React's useState and useEffect hooks for state management and event handling. The component also includes an API request handler to '/api/vector-search' and a form to submit user queries.
