+++ components/search-dialog/ui/dialog.tsx
@@ -10,9 +10,13 @@ const Dialog = DialogPrimitive.Root
 
 const DialogTrigger = DialogPrimitive.Trigger
 
-const DialogPortal = ({ className, children, ...props }: DialogPrimitive.DialogPortalProps) => (
-  <DialogPrimitive.Portal className={cn(className)} {...props}>
-    <div className="fixed inset-0 z-50 flex items-start justify-center sm:items-center">
+interface CustomDialogPortalProps extends DialogPrimitive.DialogPortalProps {
+  className?: string;
+}
+
+const DialogPortal = ({ className, children, ...props }: CustomDialogPortalProps) => (
+  <DialogPrimitive.Portal {...props}>
+    <div className={`fixed inset-0 z-50 flex items-start justify-center sm:items-center ${className}`}>
       {children}
     </div>
   </DialogPrimitive.Portal>
